feat(avatar): rename component
