--- core/components/atoms/selectionAvatarGroup/SelectionAvatarPopover/SelectionAvatarEmptyState.tsx
@@ -1,40 +1,0 @@
-import * as React from 'react';
-import { Text } from '@/index';
-
-interface EmptyStateProps {
-  /**
-   * Specify height for empty state content
-   */
-  height?: number;
-  /**
-   * Describe title for empty state
-   */
-  title?: React.ReactText;
-  /**
-   * Specify description for empty state
-   */
-  description?: string;
-}
-
-export const SelectionAvatarEmptyState = (props: EmptyStateProps) => {
-  const { height, title, description } = props;
-  return (
-    <div
-      className="d-flex flex-column justify-content-center align-items-center"
-      style={{ height: height ? height - 4 : '' }}
-    >
-      {title && (
-        <Text className="text-align-center mb-3" weight="strong">
-          {title}
-        </Text>
-      )}
-      {description && (
-        <Text className="text-align-center mb-6" weight="medium" size="small" appearance="subtle">
-          {description}
-        </Text>
-      )}
-    </div>
-  );
-};
-
-export default SelectionAvatarEmptyState;
