--- core/components/atoms/selectionAvatarGroup/SelectionAvatarPopover/index.tsx
@@ -1,5 +1,0 @@
-export * from './SelectionAvatarPopover';
-export * from './SelectionAvatarEmptyState';
-export * from './SelectionAvatarInput';
-export * from './SelectionAvatarList';
-export * from './SelectionAvatarOption';
