--- core/components/atoms/selectionAvatarGroup/SelectionAvatars/index.tsx
@@ -1,2 +1,0 @@
-export * from './SelectionAvatars';
-export * from './SelectionAvatarCount';
