--- core/components/atoms/selectionAvatarGroup/index.tsx
@@ -1,2 +1,0 @@
-export { default } from './SelectionAvatarGroup';
-export * from './SelectionAvatarGroup';
